Create Counter component
